DELTASPIKE-1369: Move SchedulerControl to spi package.
